Fixed a bug where links were not correctly saved
Fixed a bug where the hard links were not correctly saved.  Instead, the files were just added as empty files to the tar.  Have to set the GNU header correctly to be of type (hard-)link and the size to be 0.  Now the files are added correctly as hard links.